requirements.txt
PyQt5==5.15.9
google-api-python-client==2.97.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-generativeai==0.3.1
python-dotenv==1.0.0
